Backend finished
Backend finished with a simple front end. This has a login and registration system that checks for valid emails, has a minimum password length and then hashes the details using mongoose db.

Modules used:
express for handling your routes
express-session for building a user session
express-ejs-layouts layout support for ejs in Express
connect-flash will be used to display flash messages. We will cover more on this later.
passport to handle user authentication
passport-local, which is a strategy(authentication mechanism) with a username and password
mongoose for storing users in the database
bcrypt for encrypting your passwords before you store them in your database. It’s a necessity to never store passwords in plain text for obvious security reasons.
ejs module to send data from Express to ejs files