Integrated HTML
Cleaned up the HTML front end to remove some code that wasn't doing anything and then reorganized it to make it more readable. I then copied the needed code to the EJS layouts and tested them, unfortunately I couldn't get the CSS to link properly as I don't think the website was serving them, I just put the CSS in style tags to solve this.